Feat/add two new metrics (#8)
* feat: add calculations for time between occurrence and report, and victim age

Introduced new fields and calculations in TR_Carga_Fato_Crimes.ktr to compute the time between the occurrence and reporting of crimes, as well as to pass through the victim's age from the SELECT statement. This enhances the data model for better analysis of crime data.

* feat: add idempotent checks for new columns in crime fact table

Updated the SQL script to include checks for the existence of the columns `Tempo_Entre_Ocorrencia_Relato` and `Idade_Vitima` before adding them to the `Fato_Crimes` table. This ensures the script can be executed multiple times without errors, enhancing the robustness of the database setup.